Add instr mem scan chain and assoc'd test bench
Note that this doesn't work yet. There's still
problems with unknown signals in the simulation,
partly due to the changes for synthesis in DC and
partly due to how the scan chain works with the
core.

See sims/tb_riscv_imem_scan.sv to see the
attempted loading of eight instructions and the
unknown imem_wen and imem_addr values. The scan
chain seems to read inputs correctly; it just has
trouble writing to the sram through the core due
to some conflicts over imem_addr, imem_din, and
imem_en.